# api/openapi.yaml
openapi: 3.0.3
info:
  title: CortaAqui API
  version: 0.1.0
  license:
    name: Proprietário
    url: https://github.com/Mouretz/app-barber-java
  description: |
    Contrato do MVP do CortaAqui. Este arquivo vem antes de qualquer tela ou endpoint:
    o app (modo mock e modo API) e o servidor seguem o que está aqui. Mudou o contrato,
    muda este arquivo no mesmo PR.

    Regras gerais:
    - Multi-barbearia: todo recurso da casa fica embaixo de `/barbershops/{barbershopId}`.
      O servidor confere se o usuário logado tem vínculo com essa barbearia; sem vínculo
      a resposta é 404 (não revela que a barbearia existe para quem não é dela).
    - Horário: o servidor usa a hora de Brasília (`America/Sao_Paulo`). Datas e horas vão
      em ISO 8601 com offset (`2026-10-08T14:30:00-03:00`). Dia sem hora vai como `date`.
    - Dinheiro sempre em centavos, inteiro (`priceCents`). Nunca ponto flutuante.
    - Erros seguem o formato `Problem` (RFC 7807) com um `code` estável que o app usa
      para escolher a mensagem.
    - Pedido repetido é ignorado: todo POST que cria agendamento exige o cabeçalho
      `Idempotency-Key` (UUID gerado pelo app). Repetir a mesma chave devolve a mesma
      resposta da primeira vez, sem criar outro agendamento.
    - Status do agendamento só anda pra frente: `SCHEDULED` vira `COMPLETED`, `NO_SHOW`
      ou `CANCELED`, e daí não volta.
    - Sem encaixe duplo por profissional: garantido por constraint no banco. Ocupam o
...
      dá 403.
...
servers:
...
security: []
paths:
...
  /barbershops/{barbershopId}:
    parameters:
      - $ref: '#/components/parameters/BarbershopId'
    get:
      tags: [Cliente]
      operationId: getBarbershop
      summary: Página da barbearia com serviços e profissionais ativos (história 16)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/BarbershopDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /barbershops/{barbershopId}/agenda:
    parameters:
      - $ref: '#/components/parameters/BarbershopId'
    get:
      tags: [Agenda]
      operationId: getAgenda
      summary: Agenda do dia por profissional (história 2)
      description: |
        Para cada profissional: janelas de expediente do dia (fora delas a tela pinta de
        cinza), agendamentos e bloqueios. Folga de dia inteiro volta com `workingWindows`
        vazio e `timeOff` preenchido.
      security: [{ staffToken: [] }]
      parameters:
        - { name: date, in: query, required: true, schema: { type: string, format: date } }
        - name: professionalId
          in: query
          description: |
            Gerente: filtra um profissional; sem filtro, vêm todos os ativos. Profissional
            que não é gerente: sem filtro recebe só a dele; pedir outro profissional dá 403.
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Agenda' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /barbershops/{barbershopId}/blocks:
    parameters:
      - $ref: '#/components/parameters/BarbershopId'
    post:
      tags: [Agenda]
      operationId: createBlock
      summary: Bloquear um intervalo (história 2)
      description: |
        Intervalo dentro de um mesmo dia, com início e fim na grade de 30 min (30 min é o
        menor). `endAt` às 00:00 do dia seguinte conta como fim do mesmo dia; passar disso
        dá 422. Profissional só bloqueia a própria agenda; gerente, qualquer uma.
        Bloquear em cima de marcação ativa é permitido e não mexe nela: a marcação fica e
        passa a vir com `overlapsBlock: true` na agenda. O único conflito é com outro
        bloqueio do mesmo profissional (409 `SLOT_TAKEN`); encostado pode.
      security: [{ staffToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BlockRequest' }
      responses:
        '201':
          description: Bloqueado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Block' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: '`SLOT_TAKEN`: bate com outro bloqueio do mesmo profissional (marcação não conta)'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /barbershops/{barbershopId}/blocks/{blockId}:
    parameters:
      - $ref: '#/components/parameters/BarbershopId'
      - name: blockId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [Agenda]
      operationId: deleteBlock
      summary: Desbloquear
      description: |
        Profissional desbloqueia só os bloqueios da própria agenda; gerente, qualquer um.
        Não mexe em nenhuma marcação; a que estava debaixo volta a `overlapsBlock: false`
        se não bater com outro bloqueio.
      security: [{ staffToken: [] }]
      responses:
        '204': { description: Removido }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------- Equipe
  /barbershops/{barbershopId}/professionals:
    parameters:
      - $ref: '#/components/parameters/BarbershopId'
    get:
      tags: [Equipe]
      operationId: listProfessionals
      summary: Profissionais da casa, inclusive inativos (história 10, só gerente)
      security: [{ staffToken: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Professional' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Equipe]
      operationId: createProfessional
      summary: Incluir profissional (só gerente)
      description: |
        Com `email`, o servidor cria o login de profissional e devolve 201; o primeiro
        acesso e a troca de senha ficam fora deste contrato por enquanto.
      security: [{ staffToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfessionalRequest' }
      responses:
        '201':
          description: Criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Professional' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /barbershops/{barbershopId}/professionals/{professionalId}:
    parameters:
      - $ref: '#/components/parameters/BarbershopId'
      - $ref: '#/components/parameters/ProfessionalId'
    patch:
      tags: [Equipe]
      operationId: updateProfessional
      summary: Editar nome, foto ou desativar (só gerente)
      description: |
        Profissional não é apagado, só desativado. Desativar com horário futuro ativo
        devolve 409 `PROFESSIONAL_HAS_FUTURE_BOOKINGS`; o gerente cancela antes. Desativado
        some da agenda e do agendamento, mas histórico e caixa continuam.
      security: [{ staffToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProfessionalPatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Professional' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`PROFESSIONAL_HAS_FUTURE_BOOKINGS`'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /barbershops/{barbershopId}/professionals/{professionalId}/working-hours:
    parameters:
      - $ref: '#/components/parameters/BarbershopId'
      - $ref: '#/components/parameters/ProfessionalId'
    get:
      tags: [Equipe]
      operationId: getWorkingHours
      summary: Expediente semanal (história 7)
      security: [{ staffToken: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WeeklyHours' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Equipe]
      operationId: replaceWorkingHours
      summary: Trocar o expediente semanal inteiro (só gerente)
      description: |
        Não apaga agendamentos que ficaram fora do novo expediente; eles continuam na
        agenda e a tela mostra em cinza.
      security: [{ staffToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WeeklyHours' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WeeklyHours' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /barbershops/{barbershopId}/professionals/{professionalId}/time-off:
    parameters:
      - $ref: '#/components/parameters/BarbershopId'
      - $ref: '#/components/parameters/ProfessionalId'
    get:
      tags: [Equipe]
      operationId: listTimeOff
      summary: Folgas a partir de hoje (história 7)
      security: [{ staffToken: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TimeOff' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Equipe]
      operationId: createTimeOff
      summary: Lançar folga, dia inteiro ou intervalo (só gerente)
      security: [{ staffToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TimeOffRequest' }
      responses:
        '201':
          description: Criada
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimeOff' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: '`TIME_OFF_HAS_BOOKINGS`: existe agendamento ativo no período; cancele antes'
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /barbershops/{barbershopId}/professionals/{professionalId}/time-off/{timeOffId}:
    parameters:
      - $ref: '#/components/parameters/BarbershopId'
      - $ref: '#/components/parameters/ProfessionalId'
      - name: timeOffId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [Equipe]
      operationId: deleteTimeOff
      summary: Remover folga (só gerente)
      security: [{ staffToken: [] }]
      responses:
        '204': { description: Removida }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ---------------------------------------------------------------- Clientes
...
  /barbershops/{barbershopId}/services:
    parameters:
      - $ref: '#/components/parameters/BarbershopId'
    get:
      tags: [Serviços]
      operationId: listServices
      summary: Serviços da casa, inclusive inativos (história 10, só gerente)
      security: [{ staffToken: [] }]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Service' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Serviços]
      operationId: createService
      summary: Criar serviço (só gerente)
      security: [{ staffToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ServiceRequest' }
      responses:
        '201':
          description: Criado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Service' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /barbershops/{barbershopId}/services/{serviceId}:
    parameters:
      - $ref: '#/components/parameters/BarbershopId'
      - name: serviceId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      tags: [Serviços]
      operationId: updateService
      summary: Mudar preço, duração ou desativar (só gerente)
      description: Mudar o preço não altera agendamentos já criados (o preço fica gravado neles).
      security: [{ staffToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ServicePatch' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Service' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }

  # ---------------------------------------------------------------- Gerência
...
components:
  securitySchemes:
    staffToken:
      type: http
      scheme: bearer
      description: Token devolvido por `/auth/login` (gerente e profissional).
    clientCode:
      type: apiKey
      in: header
      name: X-Client-Code
      description: Código gerado pelo app (UUID v4) e guardado no aparelho. O servidor nunca devolve. Ausente, fora do formato, desconhecido ou liberado dá o mesmo 401.
  parameters:
    BarbershopId:
      name: barbershopId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    ProfessionalId:
      name: professionalId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Unauthorized:
      description: '`UNAUTHORIZED`: sem token, token vencido ou código do cliente inválido'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Forbidden:
      description: '`FORBIDDEN`: logado, mas o papel não permite (ex.: profissional mexendo em preço)'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: '`NOT_FOUND`: não existe ou não é desta barbearia'
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
...
    ValidationError:
      description: Regra de negócio ou campo inválido; ver `code` e `fields`
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    Problem:
      type: object
      required: [status, code, title]
      properties:
        type: { type: string, format: uri, default: about:blank }
        title: { type: string, example: Horário ocupado }
        status: { type: integer, example: 409 }
        code: { $ref: '#/components/schemas/ErrorCode' }
        detail: { type: string, description: Texto para log; o app mostra a mensagem dele pelo `code`. }
        fields:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string, example: phone }
              message: { type: string }
    ErrorCode:
      type: string
      enum:
        - UNAUTHORIZED
        - FORBIDDEN
        - NOT_FOUND
        - VALIDATION_ERROR
        - SLOT_TAKEN
        - TOO_SOON
        - OUTSIDE_WORKING_HOURS
        - BOOKING_LIMIT_REACHED
        - CANCEL_WINDOW_CLOSED
        - STATUS_CHANGED
        - PHONE_ON_OTHER_DEVICE
        - DEVICE_PHONE_MISMATCH
        - DEVICE_RELEASE_NOT_ALLOWED
        - PHONE_IN_USE
        - EMAIL_IN_USE
        - TIME_OFF_HAS_BOOKINGS
        - COMMISSION_SUM_INVALID
        - DATE_OUT_OF_RANGE
        - TOO_EARLY
        - SLOT_IN_PAST
        - PROFESSIONAL_HAS_FUTURE_BOOKINGS
        - IDEMPOTENCY_KEY_REUSED
...
    BarbershopSummary:
      type: object
      required: [id, name, neighborhood, city]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Barbearia Navalha }
        neighborhood: { type: string, example: Pinheiros }
        city: { type: string, example: São Paulo }
        coverUrl: { type: string, format: uri, nullable: true }
        startingPriceCents: { type: integer, minimum: 0, nullable: true, description: Menor preço ativo. }
    BarbershopDetail:
      allOf:
        - $ref: '#/components/schemas/BarbershopSummary'
        - type: object
          required: [address, services, professionals]
          properties:
            address: { type: string }
            phone: { type: string, nullable: true }
            services:
              type: array
              items: { $ref: '#/components/schemas/Service' }
            professionals:
              type: array
              items: { $ref: '#/components/schemas/ProfessionalPublic' }
    Service:
      type: object
      required: [id, name, durationMinutes, priceCents, active]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Corte degradê }
        durationMinutes: { type: integer, minimum: 5, maximum: 480, example: 40 }
        priceCents: { type: integer, minimum: 0, example: 5000 }
        active: { type: boolean }
    ServiceRequest:
      type: object
      required: [name, durationMinutes, priceCents]
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        durationMinutes: { type: integer, minimum: 5, maximum: 480 }
        priceCents: { type: integer, minimum: 0 }
    ServicePatch:
      type: object
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        durationMinutes: { type: integer, minimum: 5, maximum: 480 }
        priceCents: { type: integer, minimum: 0 }
        active: { type: boolean }
    ProfessionalPublic:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string, example: Caio }
        photoUrl: { type: string, format: uri, nullable: true }
    Professional:
      allOf:
        - $ref: '#/components/schemas/ProfessionalPublic'
        - type: object
          required: [active]
          properties:
            active: { type: boolean }
            email: { type: string, format: email, nullable: true, description: 'Login dele, se tiver.' }
    ProfessionalRequest:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        email: { type: string, format: email, maxLength: 120 }
        photoUrl: { type: string, format: uri }
    ProfessionalPatch:
      type: object
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 60 }
        photoUrl: { type: string, format: uri, nullable: true }
        active: { type: boolean }
    TimeRange:
      type: object
      required: [start, end]
      description: Hora local de Brasília, `HH:mm`. `end` maior que `start`, ou `end = 00:00`, que é meia-noite (fim do dia).
      properties:
        start: { type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d$', example: '09:00' }
        end: { type: string, pattern: '^([01]\d|2[0-3]):[0-5]\d$', example: '12:00' }
    Weekday:
      type: string
      enum: [MON, TUE, WED, THU, FRI, SAT, SUN]
    WeeklyHours:
      type: object
      required: [days]
      properties:
        days:
          type: array
          description: Dia ausente da lista é folga fixa. Até 3 janelas por dia (ex. manhã e tarde com almoço no meio), sem sobreposição.
          items:
            type: object
            required: [weekday, windows]
            properties:
              weekday: { $ref: '#/components/schemas/Weekday' }
              windows:
                type: array
                minItems: 1
                maxItems: 3
                items: { $ref: '#/components/schemas/TimeRange' }
    TimeOffRequest:
      type: object
      required: [startAt, endAt]
      properties:
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        reason: { type: string, maxLength: 120 }
    TimeOff:
      allOf:
        - type: object
          required: [id]
          properties:
            id: { type: string, format: uuid }
        - $ref: '#/components/schemas/TimeOffRequest'
...
    BookingStatus:
      type: string
      enum: [SCHEDULED, COMPLETED, NO_SHOW, CANCELED]
    BookingSource:
      type: string
      enum: [CLIENT_APP, STAFF, COUNTER]
    Booking:
      type: object
      required: [id, barbershopId, status, source, startAt, endAt, service, professional, client, priceCents, createdAt]
      properties:
        id: { type: string, format: uuid }
        barbershopId: { type: string, format: uuid }
        barbershopName: { type: string, description: Usado no app Cliente. }
        status: { $ref: '#/components/schemas/BookingStatus' }
        source: { $ref: '#/components/schemas/BookingSource' }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time, description: startAt + duração do serviço gravada na criação. }
        service:
          type: object
          required: [id, name]
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
        professional: { $ref: '#/components/schemas/ProfessionalPublic' }
        client:
          type: object
          required: [id, name]
          properties:
            id: { type: string, format: uuid }
            name: { type: string }
            phone: { type: string, pattern: '^\+55\d{10,11}$', nullable: true, description: Só aparece para a casa. Null no balcão sem telefone. }
        priceCents: { type: integer, minimum: 0, description: Gravado na criação. }
        shopPercent: { type: integer, minimum: 0, maximum: 100, nullable: true, description: Gravado ao concluir; só a casa vê. }
        professionalPercent: { type: integer, minimum: 0, maximum: 100, nullable: true, description: Gravado ao concluir; só a casa vê. }
        canCancel: { type: boolean, description: 'App Cliente: true se SCHEDULED e faltam 2h ou mais.' }
        canceledBy: { type: string, enum: [CLIENT, STAFF], nullable: true }
        note: { type: string, maxLength: 200, nullable: true }
        createdAt: { type: string, format: date-time }
        overlapsBlock:
          type: boolean
          description: |
            Só na Casa (agenda, balcão, mudança de status, ficha do cliente); o app Cliente não
            recebe. true quando a marcação está ativa (`SCHEDULED` ou `COMPLETED`) e bate com
            algum bloqueio do mesmo profissional; a agenda destaca para o profissional resolver.
            Calculado na leitura.
...
    BlockRequest:
      type: object
      required: [professionalId, startAt, endAt]
      properties:
        professionalId: { type: string, format: uuid }
        startAt: { type: string, format: date-time }
        endAt: { type: string, format: date-time }
        reason: { type: string, maxLength: 120 }
    Block:
      allOf:
        - type: object
          required: [id]
          properties:
            id: { type: string, format: uuid }
        - $ref: '#/components/schemas/BlockRequest'
    Agenda:
      type: object
      required: [date, professionals]
      properties:
        date: { type: string, format: date }
        professionals:
          type: array
          items:
            type: object
            required: [professional, workingWindows, bookings, blocks, timeOff]
            properties:
              professional: { $ref: '#/components/schemas/ProfessionalPublic' }
              workingWindows:
                type: array
                items: { $ref: '#/components/schemas/TimeRange' }
              bookings:
                type: array
                description: Todos os status, para a tela mostrar concluído, falta e cancelado.
                items: { $ref: '#/components/schemas/Booking' }
              blocks:
                type: array
                items: { $ref: '#/components/schemas/Block' }
              timeOff:
                type: array
                items: { $ref: '#/components/schemas/TimeOff' }
...